initial support to split completeness rules; refactor spectral rules
